Tests/LibWeb/Text/expected/css/transition-coordination.txt

left: 1000
right: 2000
top: 2000
